refactor(agents): split model fallback responsibilities
